[qt6] Add a "fixedPointSize" signal/slot
When "fixedPointSize" is set to true, the pointSize in the shaders
is immediately set to 0.005f and will not be modified until
fixedPointSize is set to false again. Any change done with the slider
on Meshroom's side is recorded, although not reflected in the rendering.
When fixedPointSize gets set back to false, the latest value set through
the slider is used to set pointSize and is effectively rendered.